+++ src/convert/slices.rs
@@ -479,7 +479,10 @@ impl<T: ErasableGeneric<Repr = JsValue> + WasmDescribe> VectorIntoWasmAbi for T 
 // JsValue-like slice support (Rust-to-JS only)
 // JsValue-like are repr(transparent) over u32, so &[JsValue] is a contiguous array of heap indices
 
-unsafe impl<T: ErasableGeneric> ErasableGeneric for &[T] {
+unsafe impl<T: ErasableGeneric> ErasableGeneric for &[T]
+where
+    <T as ErasableGeneric>::Repr: Sized,
+{
     type Repr = &'static [T::Repr];
 }